task_1/Tex_Rendering/requirement.txt
Pillow>=10.0.0
#sudo apt-get update
#sudo apt-get install -y \
#  latexmk texlive-latex-recommended texlive-latex-extra texlive-fonts-recommended \
#  texlive-fonts-extra texlive-science \
#  poppler-utils imagemagick

